OneDrive/Desktop/Todas las carpetas y app/aws_s3_demo.py/Proyecto final devopsssss/requirements.txt
mysql-connector-python
cryptography
colorama
